Issue #5 Only send ds18x messages if we have three consecutive readings.
This code in effort to stop "noise" readings I've been seeing as of
recent, especially random noise devices showing up.

Also here, this is my first compile using VisualStudio, was previously
using the Arduino IDE. Hope to transitiont to a combination of
VisualStudio and (commandline) makefile. Hopefully, the Makefile can be
used to easily install.