Force send of ds18x messages, and expire after 5 minutes
It has been difficult to program automations on temperatures where their
"last_updated" would not change for multiple minutes. With this code
change, our temperature sensors in HA update frequently, giving us more
confidence that, the value we see, is what the actual temperature is.

Also-added the "Expire After", in hopes that HA automations could key on
that to avoid automations (in my case, controlling the furnace.)