Use different thermometer icons, based on actual temperatures being sent.
TODO:
Actually SEND the discovery message, without data, if isValidReading()
is false. Here, we want to send the "alert" icon, but above this code,
we never get the chance to send due to the same condition.
However, if we do send on "isValidReading() == false" discovery, we need
to make sure there are at least two past readings, otherwise, we will
see 'noise devices' come through to HA again.